requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.1.0
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
natsort==8.4.0
pillow==11.1.0
werkzeug==3.1.3